+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/S7_ModInterface.lua
@@ -248,7 +248,7 @@ function S7_DynamicAction(option, switch)
         StatsConfigurator()
         StatsSynchronize()
         toConfigure = {}
-        S7_BuildConfigData(quickMenuVars.modName, Ext.JsonStringify(quickMenuVars.configData))
+        BuildConfigData(Ext.JsonStringify(quickMenuVars.configData), quickMenuVars.modUUID, quickMenuVars.modName)
     elseif quickMenuVars.selectedAction == "Clear" then
         quickMenuVars.selectedVal = quickMenuVars.defaultVal
     end